Footer: add category list
This adds `jojo2016_the_category_items()` to print out a group of
thumbnail images to categories, using the latest thumbnail from a
post in that category so it always updates.

This also moves all footer styles into a new _footer.scss partial
and updates the background/link color styles for the footer.