Add backbone powered portfolio
This adds 'assets/js/portfolio.jo' which holds all the backbone stuff
for an API powered portfolio app. A custom taxonomy template which
defers to archive-jetpack-portfolio.php makes everything hang together
from a PHP standpoint. The content-jetpack-portfolio.php file holds
both the PHP and JS template which is switched by checking the value of
a `js-tmpl` query var so we can call for the JS version of the template
at the bottom of the archive-jetpack-portfolio.php by first calling
`set_query_var( 'js-tmpl', true );`.

This also registers an API field for jetpack posts which adds all our
thumbnail data easily to the data attributes and caches the data using
transients.

Finally, this filters the value of `wp_calcualte_image_sizes()` to
match the needs of our layout on the portfolio page.