Add custom single templates for porfolio
This adds single-jetpack-portfolio.php which uses a new content template,
content-jetpack-portfolio-full.php to display the contents of the project.

This also adds a new function `jojo2016_filter_content_width()` to
function.php to modify the $content_width value on full-width pages. This
is needed to make Jetpack galleries stretch to the full width of the page
and not get constrained to the normal $content_width value.