Order projects by category and add tiles
This changes the main loop so it orders the projects by project
category by running the loop once for each project type and only
displaying posts that are in that particular term.

This also adds a `.jetpack-porfolio-tile` at the beginning of
each project type.

This does not affect the backbone views, which will need to be
updated.