Add mailchimp promo to posts
This adds `jojo2016_add_mailchimp_signup_to_content` as a new
content filter that will add a mailchimp signup form to the end of
each post.

This also abstracts the styles from the social widget container
into a reusable style named `.block` that is found in `_blocks.scss`
and moves all of the mailchimp styles out of `_widgets.scss` and
into `_mailchimp.scss`.

Fixes #5.