Footer: Fix bug where category items would not have image.
This modifies the query in `jojo2016_the_category_items()` such
that instead of using a meta query to find posts with a featured
image, we grab the last 5 posts from each category and loop
through them until we find one with a featured image and use that
image instead.

Fixes #4.